Separate Twitch emote detection/extraction logic to separate service
Also fixes indices being off by 1 or more depending on the amount of preceding surrogate pairs
