+++ CatCore/Services/Twitch/TwitchIrcService.cs
@@ -11,13 +11,11 @@ using CatCore.Models.EventArgs;
 using CatCore.Models.Shared;
 using CatCore.Models.Twitch;
 using CatCore.Models.Twitch.IRC;
-using CatCore.Models.Twitch.Media;
 using CatCore.Models.Twitch.OAuth;
 using CatCore.Services.Interfaces;
 using CatCore.Services.Twitch.Interfaces;
 using CatCore.Services.Twitch.Media;
 using Serilog;
-using Twemoji = CatCore.Emoji.Twemoji;
 
 namespace CatCore.Services.Twitch
 {
@@ -36,11 +34,11 @@ namespace CatCore.Services.Twitch
 		private readonly ILogger _logger;
 		private readonly IKittenWebSocketProvider _kittenWebSocketProvider;
 		private readonly IKittenPlatformActiveStateManager _activeStateManager;
-		private readonly IKittenSettingsService _settingsService;
 		private readonly ITwitchAuthService _twitchAuthService;
 		private readonly ITwitchChannelManagementService _twitchChannelManagementService;
 		private readonly ITwitchRoomStateTrackerService _roomStateTrackerService;
 		private readonly ITwitchUserStateTrackerService _userStateTrackerService;
+		private readonly TwitchEmoteDetectionHelper _twitchEmoteDetectionHelper;
 		private readonly TwitchMediaDataProvider _twitchMediaDataProvider;
 
 		private readonly Dictionary<string, string> _channelNameToChannelIdDictionary;
@@ -57,18 +55,18 @@ namespace CatCore.Services.Twitch
 		private CancellationTokenSource? _messageQueueProcessorCancellationTokenSource;
 		private ValidationResponse? _loggedInUser;
 
-		public TwitchIrcService(ILogger logger, IKittenWebSocketProvider kittenWebSocketProvider, IKittenPlatformActiveStateManager activeStateManager, IKittenSettingsService settingsService,
-			ITwitchAuthService twitchAuthService, ITwitchChannelManagementService twitchChannelManagementService, ITwitchRoomStateTrackerService roomStateTrackerService,
-			ITwitchUserStateTrackerService userStateTrackerService, TwitchMediaDataProvider twitchMediaDataProvider)
+		public TwitchIrcService(ILogger logger, IKittenWebSocketProvider kittenWebSocketProvider, IKittenPlatformActiveStateManager activeStateManager, ITwitchAuthService twitchAuthService,
+			ITwitchChannelManagementService twitchChannelManagementService, ITwitchRoomStateTrackerService roomStateTrackerService, ITwitchUserStateTrackerService userStateTrackerService,
+			TwitchEmoteDetectionHelper twitchEmoteDetectionHelper, TwitchMediaDataProvider twitchMediaDataProvider)
 		{
 			_logger = logger;
 			_kittenWebSocketProvider = kittenWebSocketProvider;
 			_activeStateManager = activeStateManager;
-			_settingsService = settingsService;
 			_twitchAuthService = twitchAuthService;
 			_twitchChannelManagementService = twitchChannelManagementService;
 			_roomStateTrackerService = roomStateTrackerService;
 			_userStateTrackerService = userStateTrackerService;
+			_twitchEmoteDetectionHelper = twitchEmoteDetectionHelper;
 			_twitchMediaDataProvider = twitchMediaDataProvider;
 
 			_twitchAuthService.OnCredentialsChanged += TwitchAuthServiceOnOnCredentialsChanged;
@@ -596,7 +594,9 @@ namespace CatCore.Services.Twitch
 				message = string.Empty;
 			}
 
-			var emotes = message.Length > 0 ? ExtractEmoteInfo(message, messageMeta, channelId, bits) : new List<IChatEmote>(0);
+			var emotes = message.Length > 0
+				? _twitchEmoteDetectionHelper.ExtractEmoteInfo(message, messageMeta, channelId, bits)
+				: new List<IChatEmote>(0);
 
 			OnMessageReceived?.Invoke(new TwitchMessage(
 				messageId,
@@ -611,124 +611,6 @@ namespace CatCore.Services.Twitch
 				commandType,
 				bits
 			));
-		}
-
-		// TODO: Look into moving this logic into its own class
-		private List<IChatEmote> ExtractEmoteInfo(string message, IReadOnlyDictionary<string, string>? messageMeta, string channelId, uint bits)
-		{
-			var emotes = new List<IChatEmote>();
-
-			var twitchConfig = _settingsService.Config.TwitchConfig;
-			if (twitchConfig.ParseTwitchEmotes && messageMeta != null)
-			{
-				ExtractTwitchEmotes(emotes, message, messageMeta);
-			}
-
-			if (_settingsService.Config.GlobalConfig.HandleEmojis)
-			{
-				ExtractEmojis(emotes, message);
-			}
-
-			ExtractOtherEmotes(emotes, message, channelId, twitchConfig.ParseCheermotes && bits > 0, twitchConfig.ParseBttvEmotes || twitchConfig.ParseFfzEmotes);
-
-			return emotes;
-		}
-
-		private static void ExtractTwitchEmotes(List<IChatEmote> emotes, string message, IReadOnlyDictionary<string, string> messageMeta)
-		{
-			if (!messageMeta.TryGetValue(IrcMessageTags.EMOTES, out var emotesString))
-			{
-				return;
-			}
-
-			var emoteGroup = emotesString.Split('/');
-			for (var i = 0; i < emoteGroup.Length; i++)
-			{
-				var emoteSet = emoteGroup[i].Split(':');
-				var emoteId = emoteSet[0];
-
-				var emotePlaceholders = emoteSet[1].Split(',');
-
-				for (var j = 0; j < emotePlaceholders.Length; j++)
-				{
-					var emoteMeta = emotePlaceholders[j].Split('-');
-					var emoteStart = int.Parse(emoteMeta[0]);
-					var emoteEnd = int.Parse(emoteMeta[1]);
-
-					emotes.Add(new TwitchEmote("TwitchEmote_" + emoteId, message.Substring(emoteStart, emoteEnd + 1 - emoteStart), emoteStart, emoteEnd,
-						$"https://static-cdn.jtvnw.net/emoticons/v2/{emoteId}/static/dark/3.0"));
-				}
-			}
-		}
-
-		private static void ExtractEmojis(List<IChatEmote> emotes, string message)
-		{
-			for (var i = 0; i < message.Length; i++)
-			{
-				var foundEmojiLeaf = Twemoji.Emojis.EmojiReferenceData.LookupLeaf(message, i);
-				if (foundEmojiLeaf != null)
-				{
-					emotes.Add(new Models.Shared.Emoji(foundEmojiLeaf.Key, foundEmojiLeaf.Key, i, i += foundEmojiLeaf.Depth, foundEmojiLeaf.Url));
-				}
-			}
-		}
-
-		// ReSharper disable once CognitiveComplexity
-		private void ExtractOtherEmotes(List<IChatEmote> emotes, string message, string channelId, bool parseCheermotes, bool parseCustomEmotes)
-		{
-			if (!parseCheermotes && !parseCustomEmotes)
-			{
-				return;
-			}
-
-			void ExtractOtherEmotesInternal(int messageStartIndex, int messageEndIndex)
-			{
-				var currentWordBuilder = new StringBuilder();
-				for (var i = messageStartIndex; i <= messageEndIndex; i++)
-				{
-					if (i == messageEndIndex || char.IsWhiteSpace(message[i]))
-					{
-						if (currentWordBuilder.Length <= 0)
-						{
-							continue;
-						}
-
-						var currentWord = currentWordBuilder.ToString();
-
-						if (parseCustomEmotes && _twitchMediaDataProvider.TryGetThirdPartyEmote(currentWord, channelId, out var customEmote))
-						{
-							var startIndex = i - currentWord.Length;
-							var endIndex = i - 1;
-
-							emotes.Add(new TwitchEmote(customEmote!.Id, customEmote.Name, startIndex, endIndex, customEmote.Url, customEmote.IsAnimated));
-						}
-						else if (parseCheermotes && _twitchMediaDataProvider.TryGetCheermote(currentWord, channelId, out var emoteBits, out var cheermoteData))
-						{
-							var startIndex = i - currentWord.Length;
-							var endIndex = i - 1;
-
-							emotes.Add(new TwitchEmote(cheermoteData!.Id, cheermoteData.Name, startIndex, endIndex, cheermoteData.Url, cheermoteData.IsAnimated, emoteBits, cheermoteData.Color));
-						}
-
-						currentWordBuilder.Clear();
-					}
-					else
-					{
-						currentWordBuilder.Append(message[i]);
-					}
-				}
-			}
-
-			var orderedEmotesList = emotes.OrderBy(x => x.StartIndex).ToList();
-			var loopStartIndex = 0;
-			foreach (var referenceEmote in orderedEmotesList)
-			{
-				ExtractOtherEmotesInternal(loopStartIndex, referenceEmote.StartIndex - 1);
-
-				loopStartIndex = referenceEmote.EndIndex + 2;
-			}
-
-			ExtractOtherEmotesInternal(loopStartIndex, message.Length);
 		}
 
 		// ReSharper disable once CognitiveComplexity
